+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
+## Version 0.5.2
+
+### Added
+
+* `CocoDataset.rebase` for changing the image root.
+
+
 ## Version 0.5.1